+++ contracts/boost-farming/src/events.rs
@@ -94,11 +94,29 @@ pub enum Event<'a> {
         withdraw_amount: &'a U128,
         success: bool,
     },
+    RewardWithdrawOutdatedBeneficiary {
+        owner_id: &'a AccountId,
+        farm_id: &'a String,
+        withdraw_amount: &'a U128,
+        success: bool,
+    },
     RewardWithdrawUndistributed {
         owner_id: &'a AccountId,
         farm_id: &'a String,
         withdraw_amount: &'a U128,
         success: bool,
+    },
+    ShadowSeedDeposit {
+        farmer_id: &'a AccountId,
+        seed_id: &'a String,
+        deposit_amount: &'a U128,
+        increased_power: &'a U128,
+        duration: u32,
+    },
+    ShadowSeedWithdraw {
+        farmer_id: &'a AccountId,
+        seed_id: &'a String,
+        withdraw_amount: &'a U128,
     },
 }