+++ contracts/boost-farming/src/farmer.rs
@@ -10,13 +10,16 @@ pub struct Farmer {
     /// Amounts of various reward tokens the farmer claimed.
     pub rewards: HashMap<AccountId, Balance>,
     /// Various seed tokens the farmer staked.
-    #[serde(skip_serializing)]
-    pub seeds: UnorderedMap<SeedId, FarmerSeed>,
+    #[serde(skip)]
+    pub seeds: UnorderedMap<SeedId, FarmerSeedOld>,
+    #[serde(skip)]
+    pub vseeds: UnorderedMap<SeedId, VFarmerSeed>,
 }
 
 #[derive(BorshSerialize, BorshDeserialize)]
 pub enum VFarmer {
     V0(FarmerV0),
+    V1(FarmerV1),
     Current(Farmer),
 }
 
@@ -24,6 +27,7 @@ impl From<VFarmer> for Farmer {
     fn from(v: VFarmer) -> Self {
         match v {
             VFarmer::V0(c) => c.into(),
+            VFarmer::V1(c) => c.into(),
             VFarmer::Current(c) => c,
         }
     }
@@ -42,6 +46,9 @@ impl Farmer {
             sponsor_id: sponsor_id.clone(),
             rewards: HashMap::new(),
             seeds: UnorderedMap::new(StorageKeys::FarmerSeed {
+                account_id: farmer_id.clone(),
+            }),
+            vseeds: UnorderedMap::new(StorageKeys::VFarmerSeed {
                 account_id: farmer_id.clone(),
             }),
         }
@@ -64,6 +71,33 @@ impl Farmer {
             }
         }
     }
+
+    pub fn get_seed_unwrap(&self, seed_id: &SeedId) -> FarmerSeed {
+        if let Some(seed) = self.seeds.get(seed_id) {
+            seed.into()
+        } else {
+            self.vseeds.get(seed_id).unwrap().into()
+        }
+    }
+
+    pub fn get_seed(&self, seed_id: &SeedId) -> Option<FarmerSeed> {
+        if let Some(seed) = self.seeds.get(seed_id) {
+            Some(seed.into())
+        } else {
+            self.vseeds.get(seed_id).map(|v| v.into())
+        }
+    }
+
+    pub fn remove_seed(&mut self, seed_id: &SeedId) {
+        if self.seeds.remove(seed_id).is_none() {
+            self.vseeds.remove(seed_id);
+        }
+    }
+
+    pub fn set_seed(&mut self, seed_id: &SeedId, seed: FarmerSeed) {
+        self.seeds.remove(seed_id);
+        self.vseeds.insert(seed_id, &seed.into());
+    }
 }
 
 impl Contract {
@@ -82,11 +116,10 @@ impl Contract {
         let mut claimed = HashMap::new();
 
         let mut farmer_seed: FarmerSeed = farmer
-            .seeds
-            .get(&seed.seed_id)
-            .map(|v| v.into())
+            .get_seed(&seed.seed_id)
             .unwrap_or_else(|| FarmerSeed {
                 free_amount: 0,
+                shadow_amount: 0,
                 locked_amount: 0,
                 x_locked_amount: 0,
                 unlock_timestamp: 0,
@@ -98,24 +131,32 @@ impl Contract {
         let farmer_seed_power = farmer_seed.get_seed_power();
 
         let mut new_user_rps = HashMap::new();
-        for (farm_id, VSeedFarm::Current(seed_farm)) in &seed.farms {
+        for (farm_id, vfarm) in &seed.farms {
+            let (seed_farm_rps, seed_farm_terms_reward_token, seed_farm_total_reward) = match vfarm {
+                VSeedFarm::V0(farm) => {
+                    (farm.rps, farm.terms.reward_token.clone(), farm.total_reward)
+                }
+                VSeedFarm::Current(farm) => {
+                    (farm.rps, farm.terms.reward_token.clone(), farm.total_reward)
+                }
+            };
             let farmer_rps = farmer_seed.user_rps.get(farm_id).unwrap_or(&BigDecimal::zero()).clone();
-            let diff = seed_farm.rps - farmer_rps;
+            let diff = seed_farm_rps - farmer_rps;
             let reward_amount = diff.round_down_mul_u128(farmer_seed_power);
             if reward_amount > 0 {
                 rewards.insert(
-                    seed_farm.terms.reward_token.clone(),
+                    seed_farm_terms_reward_token.clone(),
                     reward_amount
                         + rewards
-                            .get(&seed_farm.terms.reward_token)
+                            .get(&seed_farm_terms_reward_token)
                             .unwrap_or(&0_u128),
                 );
                 claimed.insert(farm_id.clone(), reward_amount);
             }
 
             // bypass non-reward
-            if seed_farm.total_reward > 0 {
-                new_user_rps.insert(farm_id.clone(), seed_farm.rps);
+            if seed_farm_total_reward > 0 {
+                new_user_rps.insert(farm_id.clone(), seed_farm_rps);
             }
         }
         farmer_seed.user_rps = new_user_rps;
@@ -132,7 +173,7 @@ impl Contract {
         farmer_seed.boost_ratios = self.gen_booster_ratios(&seed.seed_id, farmer);
         seed.total_seed_power = seed.total_seed_power + farmer_seed.get_seed_power() - prev;
 
-        farmer.seeds.insert(&seed.seed_id, &farmer_seed);
+        farmer.set_seed(&seed.seed_id, farmer_seed);
         seed.update_claimed(&claimed);
 
     }