+++ contracts/boost-farming/src/management.rs
@@ -5,6 +5,14 @@ use near_contract_standards::fungible_token::core_impl::ext_fungible_token;
 impl Contract {
 
     #[payable]
+    pub fn modify_ref_exchange_id(&mut self, ref_exchange_id: AccountId) {
+        assert_one_yocto();
+        require!(self.is_owner_or_operators(), E002_NOT_ALLOWED);
+        log!("Modify ref_exchange_id from {} to {}", self.data().ref_exchange_id, ref_exchange_id);  
+        self.data_mut().ref_exchange_id = ref_exchange_id;
+    }
+
+    #[payable]
     pub fn modify_daily_reward(&mut self, farm_id: FarmId, daily_reward: U128) {
         assert_one_yocto();
         require!(self.is_owner_or_operators(), E002_NOT_ALLOWED);
@@ -13,8 +21,15 @@ impl Contract {
         let (seed_id, _) = parse_farm_id(&farm_id);
         let mut seed = self.internal_unwrap_seed(&seed_id);
 
-        let VSeedFarm::Current(seed_farm) = seed.farms.get_mut(&farm_id).expect(E401_FARM_NOT_EXIST);
-        seed_farm.terms.daily_reward = daily_reward.0;
+        let vfarm = seed.farms.get_mut(&farm_id).expect(E401_FARM_NOT_EXIST);
+        match vfarm {
+            VSeedFarm::V0(farm) => {
+                farm.terms.daily_reward = daily_reward.0;
+            }
+            VSeedFarm::Current(farm) => {
+                farm.terms.daily_reward = daily_reward.0;
+            }
+        }
 
         self.internal_set_seed(&seed_id, seed);
     }
@@ -206,11 +221,21 @@ impl Contract {
         let (seed_id, _) = parse_farm_id(&farm_id);
         let mut seed = self.internal_unwrap_seed(&seed_id);
 
-        let VSeedFarm::Current(seed_farm) = seed.farms.get_mut(&farm_id).expect(E401_FARM_NOT_EXIST);
-        let amount = seed_farm.amount_of_beneficiary;
+        let v_farm = seed.farms.remove(&farm_id).expect(E401_FARM_NOT_EXIST);
+        let mut seed_farm = match v_farm {
+            VSeedFarm::V0(farm) => {
+                farm.into()
+            }
+            VSeedFarm::Current(farm) => {
+                farm
+            }
+        };
+
+        let amount = seed_farm.amount_of_beneficiary - seed_farm.amount_of_withdrew_beneficiary;
         require!(amount > 0, E101_INSUFFICIENT_BALANCE);
         let reward_token = seed_farm.terms.reward_token.clone();
-        seed_farm.amount_of_beneficiary = 0;
+        seed_farm.amount_of_withdrew_beneficiary = seed_farm.amount_of_beneficiary;
+        seed.farms.insert(farm_id.clone(), seed_farm.into());
         self.internal_set_seed(&seed_id, seed);
 
         ext_fungible_token::ft_transfer(
@@ -231,6 +256,45 @@ impl Contract {
     }
 
     #[payable]
+    pub fn withdraw_from_outdated_beneficiary_account(&mut self, farm_id: FarmId) -> Promise {
+        assert_one_yocto();
+        require!(self.is_owner_or_operators(), E002_NOT_ALLOWED);
+        require!(self.data().state == RunningState::Running, E004_CONTRACT_PAUSED);
+        
+        let v_farm = self.data_mut().outdated_farms.remove(&farm_id).expect(E401_FARM_NOT_EXIST);
+        let mut seed_farm = match v_farm {
+            VSeedFarm::V0(farm) => {
+                farm.into()
+            }
+            VSeedFarm::Current(farm) => {
+                farm
+            }
+        };
+
+        let amount = seed_farm.amount_of_beneficiary - seed_farm.amount_of_withdrew_beneficiary;
+        require!(amount > 0, E101_INSUFFICIENT_BALANCE);
+        let reward_token = seed_farm.terms.reward_token.clone();
+        seed_farm.amount_of_withdrew_beneficiary = seed_farm.amount_of_beneficiary;
+        self.data_mut().outdated_farms.insert(&farm_id, &seed_farm.into());
+
+        ext_fungible_token::ft_transfer(
+            self.data().owner_id.clone(),
+            amount.into(),
+            None,
+            reward_token,
+            1, // one yocto near
+            GAS_FOR_REWARD_TRANSFER,
+        )
+        .then(ext_self::callback_withdraw_outdated_beneficiary(
+            farm_id.clone(),
+            amount.into(),
+            env::current_account_id(),
+            0,
+            GAS_FOR_RESOLVE_REWARD_TRANSFER,
+        ))
+    }
+
+    #[payable]
     pub fn withdraw_from_undistributed_reward(&mut self, farm_id: FarmId, amount: U128) -> Promise {
         assert_one_yocto();
         require!(self.is_owner_or_operators(), E002_NOT_ALLOWED);
@@ -241,7 +305,16 @@ impl Contract {
         let (seed_id, _) = parse_farm_id(&farm_id);
         let mut seed = self.internal_unwrap_seed(&seed_id);
 
-        let VSeedFarm::Current(seed_farm) = seed.farms.get_mut(&farm_id).expect(E401_FARM_NOT_EXIST);
+        let v_farm = seed.farms.remove(&farm_id).expect(E401_FARM_NOT_EXIST);
+        let mut seed_farm = match v_farm {
+            VSeedFarm::V0(farm) => {
+                farm.into()
+            }
+            VSeedFarm::Current(farm) => {
+                farm
+            }
+        };
+        
         let amount_availabe = seed_farm.total_reward - seed_farm.distributed_reward;
         if amount == 0 {
             amount = amount_availabe;
@@ -249,6 +322,7 @@ impl Contract {
         require!(amount <= amount_availabe && amount_availabe > 0, E101_INSUFFICIENT_BALANCE);
         let reward_token = seed_farm.terms.reward_token.clone();
         seed_farm.total_reward -= amount;
+        seed.farms.insert(farm_id.clone(), seed_farm.into());
         self.internal_set_seed(&seed_id, seed);
 
         ext_fungible_token::ft_transfer(
@@ -352,8 +426,11 @@ impl Contract {
                 // without touching any seed distribution logic to save gas consumption
                 let (seed_id, _) = parse_farm_id(&farm_id);
                 let mut seed: Seed = self.data().seeds.get(&seed_id).map(|v| v.into()).expect(E301_SEED_NOT_EXIST);
-                let VSeedFarm::Current(seed_farm) = seed.farms.get_mut(&farm_id).expect(E401_FARM_NOT_EXIST);
-                seed_farm.amount_of_beneficiary += amount;
+                if let VSeedFarm::Current(farm) = seed.farms.get_mut(&farm_id).expect(E401_FARM_NOT_EXIST) {
+                    farm.amount_of_withdrew_beneficiary -= amount;
+                } else {
+                    env::panic_str(E006_NOT_IMPLEMENTED);
+                }
                 self.data_mut().seeds.insert(&seed_id, &seed.into());
 
                 Event::RewardWithdrawBeneficiary {
@@ -366,6 +443,43 @@ impl Contract {
             },
             PromiseResult::Successful(_) => {
                 Event::RewardWithdrawBeneficiary {
+                    owner_id: &self.data().owner_id,
+                    farm_id: &farm_id,
+                    withdraw_amount: &U128(amount),
+                    success: true,
+                }
+                .emit();
+            }
+        }
+    }
+
+    #[private]
+    pub fn callback_withdraw_outdated_beneficiary(&mut self, farm_id: FarmId, amount: U128) {
+        require!(
+            env::promise_results_count() == 1,
+            E001_PROMISE_RESULT_COUNT_INVALID
+        );
+        let amount: Balance = amount.into();
+        match env::promise_result(0) {
+            PromiseResult::NotReady => unreachable!(),
+            PromiseResult::Failed => {
+                if let VSeedFarm::Current(mut farm) = self.data_mut().outdated_farms.remove(&farm_id).expect(E401_FARM_NOT_EXIST) {
+                    farm.amount_of_withdrew_beneficiary -= amount;
+                    self.data_mut().outdated_farms.insert(&farm_id, &farm.into());
+                } else {
+                    env::panic_str(E006_NOT_IMPLEMENTED);
+                }
+
+                Event::RewardWithdrawOutdatedBeneficiary {
+                    owner_id: &self.data().owner_id,
+                    farm_id: &farm_id,
+                    withdraw_amount: &U128(amount),
+                    success: false,
+                }
+                .emit();
+            },
+            PromiseResult::Successful(_) => {
+                Event::RewardWithdrawOutdatedBeneficiary {
                     owner_id: &self.data().owner_id,
                     farm_id: &farm_id,
                     withdraw_amount: &U128(amount),
@@ -390,8 +504,12 @@ impl Contract {
                 // without touching any seed distribution logic to save gas consumption
                 let (seed_id, _) = parse_farm_id(&farm_id);
                 let mut seed: Seed = self.data().seeds.get(&seed_id).map(|v| v.into()).expect(E301_SEED_NOT_EXIST);
-                let VSeedFarm::Current(seed_farm) = seed.farms.get_mut(&farm_id).expect(E401_FARM_NOT_EXIST);
-                seed_farm.total_reward += amount;
+                if let VSeedFarm::Current(farm) = seed.farms.get_mut(&farm_id).expect(E401_FARM_NOT_EXIST) {
+                    farm.total_reward += amount;
+                } else {
+                    env::panic_str(E006_NOT_IMPLEMENTED);
+                }
+
                 self.data_mut().seeds.insert(&seed_id, &seed.into());
 
                 Event::RewardWithdrawUndistributed {